test: support display positioning
