+++ spec/fixtures/native-addon/virtual-display/lib/virtual-display.js
@@ -1,6 +1,6 @@
 module.exports = process.platform === 'darwin'
   ? require('../build/Release/virtual_display.node')
   : {
-      addDisplay: () => { throw new Error('Virtual displays only supported on macOS'); },
-      removeDisplay: () => { throw new Error('Virtual displays only supported on macOS'); }
+      create: () => { throw new Error('Virtual displays only supported on macOS'); },
+      destroy: () => { throw new Error('Virtual displays only supported on macOS'); }
     };
